overfort flere tester for AGP2 til nye endringer i regelverket
